:lipstick:  Style: 한글->영어, images, files 박스 색상 변경
Closes #119
