+++ components/domain/ContentBox/index.jsx
@@ -5,14 +5,15 @@ const ContentBox = ({ title, fontSize, children }) => {
   const Container = styled.div`
     width: 100%;
     height: 200px;
-    border: 1px solid darkgray;
+    border: 1px solid #bababa;
+
     border-radius: 10px;
     margin-top: 20px;
     margin-bottom: 20px;
   `
   const Title = styled.div`
     width: 100%;
-    background-color: #cfcfcf;
+    background-color: rgba(207, 207, 207, 0.12);
     padding: 10px 0 10px 15px;
     border-top-left-radius: 10px;
     border-top-right-radius: 10px;
