+++ components/domain/Profile/index.jsx
@@ -1,0 +1,32 @@
+import styled from '@emotion/styled'
+import { Avatar, Text } from '../../base'
+
+const Profile = ({ profilePhotoUrl, userName, email }) => {
+  const avatarArgs = {
+    src: profilePhotoUrl,
+    size: '150%',
+    alt: 'avatar',
+  }
+  const avatarStyle = {
+    marginTop: 10,
+    marginBottom: 10,
+  }
+  const Container = styled.div`
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+  `
+  return (
+    <Container>
+      <Avatar {...avatarArgs} style={avatarStyle} />
+      <Text size={25} strong>
+        {userName}
+      </Text>
+      <Text size={15} color="#8D8D8D">
+        {email}
+      </Text>
+    </Container>
+  )
+}
+export default Profile
